Moved smart switch content to different class
